+++ .firebaserc
@@ -1,5 +1,5 @@
 {
   "projects": {
-    "default": "wefriiends-cd7a3"
+    "default": "wefriiends-f8f52"
   }
 }
