--- src/components/chatExample/FirebaseConfig.tsx
@@ -1,11 +1,0 @@
-// Your web app's Firebase configuration
-export const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-  authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_FIREBASE_APP_ID,
-  databaseURL: process.env.REACT_APP_FIREBASE_DATABASE_URL,
-  measurementId: process.env.REACT_APP_FIREBASE_MEASUREMENT_ID,
-}
