+++ src/components/chatExample/firestoreChat/FirestoreChat.tsx
@@ -1,5 +1,3 @@
-import { initializeApp } from 'firebase/app'
-import { getFirestore } from 'firebase/firestore'
 import React, { useState, useEffect } from 'react'
 import {
   collection,
@@ -12,18 +10,7 @@ import {
   limitToLast,
 } from 'firebase/firestore'
 
-const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-  authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
-  storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.REACT_APP_FIREBASE_MESSAGING_SENDER_ID,
-  appId: process.env.REACT_APP_FIREBASE_APP_ID,
-  databaseURL: process.env.REACT_APP_FIREBASE_DATABASE_URL,
-}
-
-const app = initializeApp(firebaseConfig)
-const db = getFirestore(app)
+import { db } from '../firebase'
 
 interface Message {
   id: string
