+++ src/components/chatExample/realtimeDatabaseChat/RealtimeDatabaseChat.tsx
@@ -14,7 +14,7 @@ import React, { useState, useEffect } from 'react'
  */
 
 const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
+  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
   authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
   projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
   storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
