+++ src/components/chatExample/realtimeDatabaseChat/RealtimeDatabaseChatStyled.tsx
@@ -12,10 +12,9 @@ import {
 import { makeStyles } from 'tss-react/mui'
 import theme from '../../../styles/createTheme'
 import ChatMenu from '../../chat/ChatMenu'
-//import { firebaseConfig } from './FirebaseConfig'
 
 const firebaseConfig = {
-  apiKey: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
+  apiKey: process.env.REACT_APP_FIREBASE_API_KEY,
   authDomain: process.env.REACT_APP_FIREBASE_AUTH_DOMAIN,
   projectId: process.env.REACT_APP_FIREBASE_PROJECT_ID,
   storageBucket: process.env.REACT_APP_FIREBASE_STORAGE_BUCKET,
